test(web-builder): move tests of plugins to `tests/plugins`, add tests for define plugin
